RFC-004 P6: handler wave 3 — state, comms, validation domains
The final 8 handlers leave ToolExecutor:
- state.py: memory_manage + manage_list with their persistence helpers
  (_lists_path/_load_lists/_save_lists/_format_list/_manage_list_locked)
  and MEMORY_MAX_KEYS_PER_SECTION (sole consumer; its two monkeypatch
  test sites re-pointed). The LOCK-HEAVY domain, isolated per plan
  advisory #2: locks + memory path stay ON the executor, reached live
  through deps. Memory PERSISTENCE primitives (_load_all_memory /
  _save_all_memory / _load_memory / _load_memory_for_user) STAY CORE —
  prompts.py and the web API call them on the executor directly.
- comms.py: email_send/search/read/list_recent + _email_cfg +
  issue_tracker (client reached live via getattr accessor, preserving
  the hasattr semantics)
- validation.py: validate_action (governor via live accessor)

Deps grows 6 state accessors + 2 memory passthroughs. Mechanical
adjustments: two-dot relative imports re-anchored to three dots
(..notifications/..llm); comms was missing an asyncio import the
covered paths never hit — added; one pre-existing dead lazy datetime
import absorbed.

3 more __new__-bypass fixtures constructed properly; email fallback +
direct-call sites re-pointed to domain owners.

executor.py 1,298 -> 718 lines (from 1,893 at campaign start — the
middleware core: execute pipeline, recovery, SSH plumbing, RBAC/risk,
memory persistence). Full suite 6,125 passed; lint gate new=0.
